// src/core/types/entities/iron-lake/media-library/ListItem.ts
export type ListItem = {
  AttachmentId: string
  Code: string
  OriginalFileName: string
  FormattedFileName: string
  FileType: string
  FileUrl: string
  Description: string
  CreatedOn: string
  CreatedBy: string
  ChangedOn: string
  ChangedBy: string
  IsActive: boolean
  isDeleted: boolean
};